refactor(web): extract SURFACE_CLASS to dedicated module
Co-located constant and component exports broke React Fast Refresh.
Move `SURFACE_CLASS` / `SurfaceVariant` into `surface-classes.ts` so
`surface.tsx` only exports the component. ProfileMenu and
PushNotificationButton updated to import from the new module.
